requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
dhooks==1.1.4
discord.py==1.7.2
greenlet==1.1.0
idna==2.10
iniconfig==1.1.1
multidict==5.1.0
packaging==20.9
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.7
pytest==6.2.4
python-dotenv==0.17.1
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.18
dnspython==1.16.0
greenlet==1.1.0
idna==2.10
iniconfig==1.1.1
mongoengine==0.23.1
multidict==5.1.0
packaging==20.9
pluggy==0.13.1
py==1.10.0
pycparser==2.20
pymongo==3.12.0
PyNaCl==1.4.0
pyparsing==2.4.7
pytest==6.2.4
pytest-asyncio==0.15.1
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
six==1.16.0
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.5
yarl==1.6.3
